+++ comicbox/__init__.py
@@ -3,6 +3,6 @@
 from os import environ
 
 if environ.get("PYTHONDEVMODE"):
-    from icecream import install
+    from icecream import install  # pyright: ignore[reportPrivateImportUsage]
 
     install()