+++ comicbox/schemas/metroninfo/identifiers.py
@@ -4,7 +4,7 @@ from types import MappingProxyType
 
 from comicbox.fields.fields import StringField
 from comicbox.fields.metroninfo import MetronIDAttrField, MetronSourceField
-from comicbox.fields.xml_fields import XmlBooleanAttributeField, XmlStringField
+from comicbox.fields.xml_fields import XmlBooleanField, XmlStringField
 from comicbox.schemas.xml_schemas import XmlSubSchema
 
 
@@ -28,7 +28,7 @@ class MetronPrimaryAttrSchema(XmlSubSchema):
         include = MappingProxyType(
             {
                 "#text": StringField(required=True),
-                "@primary": XmlBooleanAttributeField(),
+                "@primary": XmlBooleanField(),
             }
         )