+++ comicbox/schemas/metroninfo/price.py
@@ -4,7 +4,7 @@ from decimal import Decimal
 from types import MappingProxyType
 
 from comicbox.fields.pycountry import CountryField
-from comicbox.fields.xml_fields import XmlTextDecimalField
+from comicbox.fields.xml_fields import XmlDecimalField
 from comicbox.schemas.xml_schemas import XmlSubSchema
 
 
@@ -16,9 +16,7 @@ class MetronPriceSchema(XmlSubSchema):
 
         include = MappingProxyType(
             {
-                "#text": XmlTextDecimalField(
-                    required=True, places=2, minimum=Decimal(0)
-                ),
+                "#text": XmlDecimalField(required=True, places=2, minimum=Decimal(0)),
                 "@country": CountryField(),
             }
         )